requirements.txt
openai==1.3.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
pygame==2.5.2
requests==2.31.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.0
pycaw==20230407
psutil==5.9.6
keyboard==0.13.5
pyautogui==0.9.54
threading
json
re
os
subprocess
platform
time
logging
